+++ mvc/src/main/java/nextstep/mvc/exception/FailMapHandler.java
@@ -1,8 +1,8 @@
-package nextstep.mvc.controller.tobe.exception;
+package nextstep.mvc.exception;
 
 public class FailMapHandler extends RuntimeException {
 
-    public FailMapHandler(){
+    public FailMapHandler() {
         super("매핑 중 오류가 발생하였습니다.");
     }
 }
